pyproject.toml
[project]
name = "tuner"
version = "0.1.0"
description = "An application to find other Spotify users with similar music tastes"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "onnx>=1.17.0",
    "onnxruntime>=1.20.0",
    "pinecone[grpc]>=5.3.1",
    "python-dotenv>=1.0.1",
    "spotipy>=2.24.0",
    "transformers>=4.46.2",
]

[project.scripts]
"tuner" = "tuner.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["tuner"]

[dependency-groups]
dev = [
    "sentence-transformers>=3.3.0",
    "scikit-learn>=1.5.2",
    "ruff>=0.7.3",
    "pytest>=8.3.3",
    "ipython>=8.29.0",
]
